requirements.txt
tensorboard
librosa
tqdm
transformers
efficientnet_pytorch
seaborn
matplotlib